feat: 상품 생성시 새 Product 엔티티 적용, 테스트 코드 수정, ProductService 요청객체 생성

+++ src/main/java/kitchenpos/products/tobe/domain/Price.java
@@ -3,27 +3,29 @@ package kitchenpos.products.tobe.domain;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
+import java.math.BigDecimal;
+
 import static kitchenpos.products.exception.ProductExceptionMessage.PRODUCT_PRICE_MORE_ZERO;
 
 @Embeddable
 public class Price {
     @Column(name = "price", nullable = false)
-    private Integer price;
+    private BigDecimal price;
 
     protected Price() {}
 
-    private Price(Integer price) {
-        if (price < 0) {
+    private Price(BigDecimal price) {
+        if (price.intValue() < 0) {
             throw new IllegalArgumentException(PRODUCT_PRICE_MORE_ZERO);
         }
         this.price = price;
     }
 
-    public static Price of(Integer price) {
+    public static Price of(BigDecimal price) {
         return new Price(price);
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
